backend/requirements-dev.txt
# Development and Testing Dependencies
-r requirements.txt

# Pre-commit Hooks
pre-commit==3.6.0

# Code Quality
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-bugbear==24.1.17
black==24.1.1
isort==5.13.2

# Django Tools
django-upgrade==1.16.0

# Security Scanning
bandit==1.7.6
safety==3.0.1

# Type Checking
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Enhanced Testing
pytest-xdist==3.5.0  # Parallel test execution
pytest-timeout==2.2.0  # Test timeouts
freezegun==1.5.1  # Already in main requirements but worth noting
